+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changed
   * AZ-1310: Update default `load_balancer_type` of Host Pool resource (`DepthFirst` => `BreadthFirst`)
   * AZ-1310: Change in method of retrieving AVD Service Principal client ID
   * AZ-1310: Change `var.scaling_plan_config.role_assignment.principal_id` to `var.scaling_plan_config.role_assignment.object_id`
+  * AZ-1310: Update example
 
 # v7.2.1 - 2024-01-19